# openapi.yaml
openapi: 3.1.0
info:
  title: MinusPod API
  description: |
    REST API for managing podcast feeds with automatic ad detection and removal.


    This API allows you to:
    - Manage podcast feed subscriptions
    - View and manage processed episodes
    - Configure ad detection settings including AI model selection
    - Monitor system status and trigger cleanup operations
    - Manage cross-episode ad patterns with network and podcast scope
    - Submit corrections to improve ad detection accuracy
  version: 2.43.1
  contact:
    name: MinusPod
  license:
    name: MIT
servers:
  # Same-origin relative URL. Swagger UI served from /api/v1/docs uses
  # whatever host the browser is on, so "Try it out" always hits the
  # correct server: localhost in dev, the production domain in prod,
  # a Tailscale/Cloudflare URL on a tunnel, etc. No hardcoded host.
  - url: /api/v1
    description: Same-origin (current server)
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: '{baseUrl}/api/v1'
    description: Custom server
    variables:
      baseUrl:
        default: http://localhost:8000
        description: Base URL of the podcast server
...
paths:
...
  /cue-detections/{detectionId}/verdict:
    post:
      tags: [AudioCueTemplates]
      summary: Confirm, reject, or reset a cue detection's review verdict
      description: |
        Advisory template-quality signal only. A verdict never changes an
        episode's cut list; it records whether a template cue matched a real ad
        boundary so a feed's cues can be judged and thresholds tuned.
      operationId: setCueDetectionVerdict
      parameters:
        - name: detectionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [verdict]
              properties:
                verdict:
                  type: string
                  enum: [pending, confirmed, rejected]
      responses:
        '200':
          description: Updated verdict
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  verdict:
                    type: string
                    enum: [pending, confirmed, rejected]
        '400':
          description: Invalid verdict
        '404':
          description: Cue detection not found
  /feeds/{slug}/cue-detections/advisory:
    get:
      tags: [AudioCueTemplates]
      summary: Per-feed cue health summary
      description: Outcome and verdict counts, match-score range, and confirm rate across the feed's processed episodes.
      operationId: getCueFeedAdvisory
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Per-feed cue advisory
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CueAdvisory'
                  - type: object
                    properties:
                      templateHints:
                        type: array
                        description: Per-template verdict hints. Only templates with enough rejections to classify are included.
                        items:
                          type: object
                          required: [templateId, hint, rejected, confirmed]
                          properties:
                            templateId:
                              type: integer
                            label:
                              type: string
                              nullable: true
                            hint:
                              type: string
                              enum: [raise_threshold, recapture]
                              description: raise_threshold when rejections cluster just above the threshold; recapture when they span the score range or overlap confirmed scores.
                            rejected:
                              type: integer
                              description: Number of rejected detections for this template.
                            confirmed:
                              type: integer
                              description: Number of confirmed detections for this template.
        '404':
          description: Feed not found
...
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: |
        Flask session cookie issued by ``POST /auth/login``. Every mutating
        endpoint additionally requires a matching ``X-CSRF-Token`` header
        (double-submit). Public endpoints (``/auth/login``, ``/auth/logout``,
        ``/auth/status``, ``/health``, ``/health/live``) declare ``security: []``.
...
  schemas:
...
    CueAdvisory:
      type: object
      description: >
        Aggregate cue telemetry for a feed (or globally). Totals, score range,
        and confirmRate count above-threshold cues only; below_threshold
        near-misses are reported separately (see the aggregate endpoint's
        nearMissTotal / nearMissHistogram).
      properties:
        total:
          type: integer
        snapped:
          type: integer
        paired:
          type: integer
        unused:
          type: integer
        confirmed:
          type: integer
        rejected:
          type: integer
        pending:
          type: integer
        avgScore:
          type: number
          nullable: true
        minScore:
          type: number
          nullable: true
        maxScore:
          type: number
          nullable: true
        confirmRate:
          type: number
          nullable: true
          description: confirmed / (confirmed + rejected), or null if none reviewed
...